requirements.txt
# WSI Tiler Requirements
# Python package for reading whole-slide images
openslide-python>=1.3.0

